- ChipAuthentication supports AES. - PACE-GM uses keyagreement that keeps the Y-coordinate, to be used in authentication token (reconstructing using curve equation apparently has a bug). - Use BigInteger for keyID in ChipAuthentication related SecurityInfos, allowing negative valued keyIDs. - First implementation of PACE-CAM. - Other first lessons learned after SDW2016 interop.
git-svn-id: https://svn.code.sf.net/p/jmrtd/code/trunk@1621 32d68b02-0bac-436f-a684-189869639a7e

+++ jmrtd/src/main/java/org/jmrtd/lds/LDSFileUtil.java
@@ -260,7 +260,7 @@ public class LDSFileUtil {
    * 
    * @return a data group number
    */
-  public static short lookupDataGroupNumberByFID(short fid) {
+  public static int lookupDataGroupNumberByFID(short fid) {
     switch(fid) {
       case PassportService.EF_DG1: return 1;
       case PassportService.EF_DG2: return 2;
